Use object instead of switch cases for color codes
